refactor: lint man. instant elapsed and unch. dur. subtr. in single pass
